Refactor: Implementación de patrón Strategy para manejo fiscal por país
- Reorganización de la lógica fiscal en clases de estrategia específicas para cada país.
- Mejora en la escalabilidad para futuros países con impuestos diferentes.
- Simplificación de la clase de servicio
